+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 
+
+### New Functions 🚀
+
+- `use_service_worker`
+
 ## [0.7.1] - 2023-10-02 
 
 ### New Function 🚀